refactor: Rename app folder to store
