+++ frontend/src/App.tsx
@@ -1,6 +1,6 @@
 import React, { useEffect } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
-import { AppDispatch, RootState } from './app/store';
+import { AppDispatch, RootState } from './store/store';
 import { fetchCurrentUser, fetchUsers } from './features/usersSlice';
 import { fetchPolls } from './features/pollSlice';
 import { getToken } from './server/api';
