+++ frontend/src/components/Leaderboard.tsx
@@ -2,7 +2,7 @@ import { createSelector } from 'reselect';
 import { useEffect } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
 import { fetchUsers } from '../features/usersSlice';
-import { AppDispatch, RootState } from '../app/store';
+import { AppDispatch, RootState } from '../store/store';
 import { Link } from 'react-router-dom';
 import {
   List,
