+++ frontend/src/components/Polls/Poll.tsx
@@ -1,7 +1,7 @@
 import React, { useEffect } from 'react';
 import { Link as RouterLink } from 'react-router-dom';
 import { useSelector, useDispatch } from 'react-redux';
-import { RootState, AppDispatch } from '../../app/store';
+import { RootState, AppDispatch } from '../../store/store';
 import { fetchPolls } from '../../features/pollSlice'; // Action to fetch polls from the server
 import { fetchUsers } from '../../features/usersSlice'; // Action to fetch users from the server
 import {
