+++ frontend/src/index.tsx
@@ -3,7 +3,7 @@ import ReactDOM from 'react-dom/client';
 import App from './App';
 import { BrowserRouter } from 'react-router-dom';
 import { Provider } from 'react-redux';
-import { store } from './app/store';
+import { store } from './store/store';
 import { ThemeProvider, CssBaseline } from '@mui/material';
 import theme from './theme';
 import 'normalize.css';
